improve potability of pid

+++ exp/pid/src/pid.c
@@ -7,7 +7,7 @@ void pid_init(float DELTA_T, float KP, float KI, float KD) {
     kd = KD;
 }
 
-float pid_calc(int16_t result, int16_t target, float error[2], float *integral) {
+float pid_calc(float result, float target, float error[2], float *integral) {
     float p, i, d;
 
     error[0] = error[1];
