+++ exp/pid/src/pid.h
@@ -6,4 +6,4 @@ static float ki = 1;
 static float kd = 1;
 
 void pid_init(float DELTA_T, float KP, float KI, float KD);
-float pid_calc(int16_t result, int16_t target, float error[2], float *integral);
+float pid_calc(float result, float target, float error[2], float *integral);
